Add serviceloader driven codec
