+++ logger-codec-lz4/pom.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
 <!--
   ~ Copyright (C) 2014-2017  Chronicle Software
   ~
@@ -18,6 +17,7 @@
   ~ limitations under the License.
   ~
   -->
+
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
@@ -29,28 +29,28 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
+    <name>OpenHFT/Chronicle-Logger/logger-codec-lz4</name>
+    <artifactId>chronicle-logger-codec-lz4</artifactId>
     <packaging>bundle</packaging>
-    <name>OpenHFT/Chronicle-Logger/logger</name>
-    <artifactId>chronicle-logger-core</artifactId>
 
     <dependencies>
         <dependency>
             <groupId>net.openhft</groupId>
-            <artifactId>chronicle-wire</artifactId>
-        </dependency>
-        
-        <!-- https://mvnrepository.com/artifact/com.google.flatbuffers/flatbuffers-java -->
-        <dependency>
-            <groupId>com.google.flatbuffers</groupId>
-            <artifactId>flatbuffers-java</artifactId>
-            <version>1.12.0</version>
+            <artifactId>chronicle-logger-core</artifactId>
         </dependency>
 
-        <!-- https://mvnrepository.com/artifact/com.moandjiezana.toml/toml4j -->
         <dependency>
-            <groupId>com.moandjiezana.toml</groupId>
-            <artifactId>toml4j</artifactId>
-            <version>0.7.2</version>
+            <groupId>org.lz4</groupId>
+            <artifactId>lz4-java</artifactId>
+            <version>1.7.1</version>
+          </dependency>
+
+        <!-- for tests -->
+        <dependency>
+            <groupId>net.openhft</groupId>
+            <artifactId>chronicle-logger-logback</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
 
@@ -64,6 +64,15 @@
                     <source>1.8</source>
                     <target>1.8</target>
                     <encoding>UTF-8</encoding>
+                    <annotationProcessorPaths>
+                        <!-- need a better jdk11 solution -->
+                        <!-- otherwise this will only work with jdk 1.8 -->
+                        <path>
+                            <groupId>jakarta.annotation</groupId>
+                            <artifactId>jakarta.annotation-api</artifactId>
+                            <version>1.3.5</version>
+                        </path>
+                    </annotationProcessorPaths>
                 </configuration>
             </plugin>
             <plugin>
@@ -72,13 +81,12 @@
                 <configuration>
                     <forkCount>0</forkCount>
                     <reuseForks>false</reuseForks>
-                    <systemPropertyVariables>
-                        <project.build.directory>${project.build.directory}</project.build.directory>
-                    </systemPropertyVariables>
-
                     <excludes>
                         <exclude>**/*PerfTest.java</exclude>
                     </excludes>
+                    <systemPropertyVariables>
+                        <project.build.directory>${project.build.directory}</project.build.directory>
+                    </systemPropertyVariables>
                 </configuration>
             </plugin>
             <!--
@@ -107,10 +115,10 @@
                 <extensions>true</extensions>
                 <configuration>
                     <instructions>
-                        <Bundle-SymbolicName>${project.groupId}.${project.artifactId}
-                        </Bundle-SymbolicName>
-                        <Bundle-Name>${project.artifactId}</Bundle-Name>
-                        <Export-Package>net.openhft.chronicle.logger.*</Export-Package>
+                        <Bundle-SymbolicName>${project.groupId}.${project.artifactId}</Bundle-SymbolicName>
+                        <Bundle-Name>OpenHFT :: ${project.artifactId}</Bundle-Name>
+                        <Bundle-Version>${project.version}</Bundle-Version>
+                        <Export-Package>net.openhft.chronicle.logger.tools.*</Export-Package>
                     </instructions>
                 </configuration>
                 <executions>
@@ -127,5 +135,12 @@
             </plugin>
         </plugins>
     </build>
+
+    <scm>
+        <url>scm:git:git@github.com:OpenHFT/Chronicle-Logger.git</url>
+        <connection>scm:git:git@github.com:OpenHFT/Chronicle-Logger.git</connection>
+        <developerConnection>scm:git:git@github.com:OpenHFT/Chronicle-Logger.git</developerConnection>
+        <tag>master</tag>
+    </scm>
 
 </project>
