+++ logger-codec-zstd/src/main/resources/META-INF/services/net.openhft.chronicle.logger.codec.CodecFactory
@@ -1,0 +1,1 @@
+net.openhft.chronicle.logger.codec.zstd.ZStandardCodecFactory
