+++ logger-core/src/main/java/net/openhft/chronicle/logger/codec/CodecFactory.java
@@ -1,0 +1,10 @@
+package net.openhft.chronicle.logger.codec;
+
+public interface CodecFactory extends AutoCloseable {
+    Codec find(CodecRegistry registry, String encodingName);
+
+    @Override
+    default void close() {
+
+    }
+}
