+++ logger-core/src/main/java/net/openhft/chronicle/logger/codec/CodecRegistry.java
@@ -1,91 +1,64 @@
 package net.openhft.chronicle.logger.codec;
 
-import net.openhft.chronicle.bytes.Bytes;
-
 import java.io.Closeable;
-import java.nio.ByteBuffer;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Duration;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.ServiceLoader;
 
-// XXX this should use a ServiceLoader map
 public class CodecRegistry implements Closeable {
 
     public static final String IDENTITY = "identity";
-    public static final String ZSTANDARD = "zstd";
 
-    private final HashMap<String, Codec> codecMap;
+    private final Duration initialDelay;
+    private final Path path;
 
-    CodecRegistry() {
-        codecMap = new HashMap<>();
-        codecMap.put(IDENTITY, new IdentityCodec());
-    }
-
-    public void addCodec(String key, Codec codec) {
-        codecMap.put(key, codec);
+    CodecRegistry(Path path, Duration initialDelay) {
+        this.path = path;
+        this.initialDelay = initialDelay;
     }
 
     public Codec find(String encoding) throws CodecException {
         if (encoding == null) {
             throw new IllegalArgumentException("Null encoding!");
         }
-        Codec codec = codecMap.get(encoding);
-        if (codec == null) {
-            throw new CodecException("No codec found for encoding " + encoding);
+
+        if (IDENTITY.equalsIgnoreCase(encoding)) {
+            return new IdentityCodec();
         }
-        return codec;
+
+        ServiceLoader<CodecFactory> loader = ServiceLoader.load(CodecFactory.class);
+        for (CodecFactory factory: loader) {
+            Codec codec = factory.find(this, encoding);
+            if (codec != null) {
+                return codec;
+            }
+            break;
+        }
+        throw new CodecException("No codec found for encoding " + encoding);
     }
 
     @Override
     public void close() {
-        for (Map.Entry<String, Codec> entry : codecMap.entrySet()) {
-            Codec codec = entry.getValue();
-            codec.close();
+        ServiceLoader<CodecFactory> loader = ServiceLoader.load(CodecFactory.class);
+        for (CodecFactory factory : loader) {
+            factory.close();
         }
     }
 
-    public static class IdentityCodec implements Codec {
+    public Path getPath() {
+        return path;
+    }
 
-        @Override
-        public int compress(Bytes<ByteBuffer> src, Bytes<ByteBuffer> dst) throws CodecException {
-            dst.write(src);
-            return (int) dst.readLimit();
-        }
-
-        @Override
-        public int decompress(Bytes<ByteBuffer> src, Bytes<ByteBuffer> dst) throws CodecException {
-            dst.write(src);
-            return (int) dst.readLimit();
-        }
-
-        @Override
-        public long compressBounds(int length) {
-            return length;
-        }
-
-        @Override
-        public long uncompressedSize(ByteBuffer buf) {
-            return buf.limit();
-        }
+    public Duration getInitialDelay() {
+        return initialDelay;
     }
 
     public static class Builder {
-        private final CodecRegistry registry;
         private Duration initialDelay = Duration.ZERO;
+        private Path path;
 
         Builder() {
-            registry = new CodecRegistry();
-        }
-
-        public Builder withCodec(String key, Codec codec) {
-            registry.addCodec(key, codec);
-            return this;
-        }
-
-        public CodecRegistry build() {
-            return registry;
         }
 
         public Builder withDefaults(String path) {
@@ -93,16 +66,17 @@ public class CodecRegistry implements Closeable {
         }
 
         public Builder withDefaults(Path path) {
-            ZStandardCodec codec = ZStandardCodec.builder()
-                    .withInitialDelay(initialDelay)
-                    .withDefaults(path)
-                    .build();
-            return withCodec(ZSTANDARD, codec);
+            this.path = path;
+            return this;
         }
 
         public Builder withInitialDelay(Duration initialDelay) {
             this.initialDelay = initialDelay;
             return this;
+        }
+
+        public CodecRegistry build() {
+            return new CodecRegistry(path, initialDelay);
         }
     }
