+++ logger-log4j-1/src/test/java/net/openhft/chronicle/logger/log4j1/Log4j1ChronicleLogTest.java
@@ -19,18 +19,17 @@ package net.openhft.chronicle.logger.log4j1;
 
 import net.openhft.chronicle.bytes.Bytes;
 import net.openhft.chronicle.core.io.IOTools;
-import net.openhft.chronicle.logger.EntryProcessor;
+import net.openhft.chronicle.logger.EntryTransformer;
 import net.openhft.chronicle.logger.LogAppenderConfig;
 import net.openhft.chronicle.logger.codec.Codec;
 import net.openhft.chronicle.logger.entry.EntryReader;
-import net.openhft.chronicle.logger.DefaultEntryProcessor;
+import net.openhft.chronicle.logger.DefaultEntryTransformer;
 import net.openhft.chronicle.logger.codec.CodecRegistry;
 import net.openhft.chronicle.logger.entry.Entry;
 import net.openhft.chronicle.queue.ChronicleQueue;
 import net.openhft.chronicle.queue.ExcerptTailer;
 import net.openhft.chronicle.wire.DocumentContext;
 import net.openhft.chronicle.wire.Wire;
-import net.openhft.chronicle.wire.WireType;
 import org.apache.log4j.Level;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
@@ -78,7 +77,7 @@ public class Log4j1ChronicleLogTest extends Log4j1TestBase {
             Path parent = Paths.get(cq.fileAbsolutePath()).getParent();
             CodecRegistry registry = CodecRegistry.builder().withDefaults(parent).build();
             Codec codec = registry.find(config.contentEncoding);
-            EntryProcessor<String> processor = new DefaultEntryProcessor(codec);
+            EntryTransformer<String> processor = new DefaultEntryTransformer(codec);
             EntryReader entryReader = new EntryReader();
 
             ExcerptTailer tailer = cq.createTailer();
@@ -120,7 +119,7 @@ public class Log4j1ChronicleLogTest extends Log4j1TestBase {
             Path parent = Paths.get(cq.fileAbsolutePath()).getParent();
             CodecRegistry registry = CodecRegistry.builder().withDefaults(path).build();
             Codec codec = registry.find(config.contentEncoding);
-            EntryProcessor<String> processor = new DefaultEntryProcessor(codec);
+            EntryTransformer<String> processor = new DefaultEntryTransformer(codec);
 
             ExcerptTailer tailer = cq.createTailer();
             EntryReader entryReader = new EntryReader();
