+++ logger-tools/pom.xml
@@ -39,6 +39,18 @@
             <artifactId>chronicle-logger-core</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>net.openhft</groupId>
+            <artifactId>chronicle-logger-codec-lz4</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>net.openhft</groupId>
+            <artifactId>chronicle-logger-codec-zstd</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
         <!-- slf4j -->
         <dependency>
             <groupId>org.slf4j</groupId>
